add support preview image small size

+++ src/modules/upload/upload.service.ts
@@ -15,27 +15,82 @@ class Processes {
       .resize(512, 512, { fit: 'cover', position: 'center'})
       .jpeg({ quality: 85 })
       .toBuffer();
-  }   
+  }
   public static async ICON(buffer: Buffer) {
+      return await sharp(buffer)
+        .resize(200, 200, { fit: 'cover', position: 'center'})
+        .jpeg({ quality: 85 })
+        .toBuffer();
+    }
+  public static async BANNER(buffer: Buffer) {
     return await sharp(buffer)
-      .resize(200, 200, { fit: 'cover', position: 'center'})
+      .resize(1024, 512, { fit: 'cover', position: 'center'})
       .jpeg({ quality: 85 })
       .toBuffer();
   }
-  public static async BANNER(buffer: Buffer) {
-  return await sharp(buffer)
-    .resize(1024, 512, { fit: 'cover', position: 'center'})
-    .jpeg({ quality: 85 })
-    .toBuffer();
-}
   public static async PRODUCT_IMAGE(buffer: Buffer) {
-  return await sharp(buffer)
-    .resize(800, 800, { fit: 'cover', position: 'center'})
-    .jpeg({ quality: 85 })
-    .toBuffer();
-} 
+    return await sharp(buffer)
+      // .resize(800, 800, { fit: 'cover', position: 'center'})
+      .jpeg({ quality: 85 })
+      .toBuffer();
+  } 
+  public static async PRODUCT_IMAGE_PREVIEW(buffer: Buffer) {
+    return await sharp(buffer)
+      .resize(800, 800, { fit: 'cover', position: 'center'})
+      .jpeg({ quality: 20 })
+      .toBuffer();
+  }
 }
 
+export const reProcess: (UDPath: string, processType: Upload.ProcessType, force?: boolean)=> Promise<void> = async (UDPath, processType, force = true)=> {
+  const fullPathFile = path.join(UD, UDPath)
+  const isExist = fs.existsSync(fullPathFile)
+  if (! isExist && force) {
+    throw new SystemError('reProcess: File not found')
+  }
+
+  if (isExist) {
+    const FileBuffer = await fs.promises.readFile(fullPathFile)
+    const updatedBuffer = await Processes[processType](FileBuffer)
+    await fs.promises.rm(fullPathFile)
+    await fs.promises.writeFile(fullPathFile, updatedBuffer)
+  }
+}
+
+export const copyFile: (UDSrc: string, UDDist: string, newName?: string, force?: boolean)=> Promise<void> = async(UDSrc, UDDist, newName, force = true)=> {
+  const fullPathFile = path.join(UD, UDSrc)
+  const isExist = await fs.existsSync(fullPathFile)
+
+  if (! isExist && force) {
+    throw new SystemError('copyFile: File not found')
+  }
+
+  if (newName) {
+    const splits = UDDist.split('/').reverse()
+    splits[0] = newName.includes('.') ? newName : newName + path.extname(splits[0]!)
+    UDDist = splits.reverse().join('/')
+  }
+
+  if (isExist) {
+    await fs.promises.cp(
+      fullPathFile,
+      path.join(UD, UDDist)
+    )
+  }
+}
+
+export const removeFile: (UDPath: string, force?: boolean)=> Promise<void> = async(UDPath, force = true)=> {
+  const fullPathFile = path.join(UD, UDPath)
+  const isExist = await fs.existsSync(fullPathFile)
+
+  if (! isExist && force) {
+    throw new SystemError('removeFile: File not found')
+  }
+
+  if (isExist) {
+    await fs.promises.rm(fullPathFile)
+  }
+}
 
 export const declareFile: (file: Express.Multer.File, user: User, process: Upload.ProcessType)=> Promise<Upload> = async(file, user, process)=> {
   const newFileSchema: Upload.Declare = {
@@ -75,6 +130,8 @@ export default class SaveFile {
   declare private _staticFile: string;
   declare private _processType: string;
   declare private _requiredSavedResults: number;
+  private _ignoreOldProcessType: boolean = false;
+  private _listIgnoreRemove: number[] = [];
   private _saveFilesIds: ID[]         = [];
   private _saveFilesPaths: string[]   = [];
   private _removeFilesIds: ID[]       = [];
@@ -132,6 +189,19 @@ export default class SaveFile {
     return this;
   }
 
+  ignoreOldProcessType() {
+    this._ignoreOldProcessType = true;
+    return this;
+  }
+  ignoreRemoveIndexes(indexes: number[], toIndex?: number) {
+    this._listIgnoreRemove = indexes;
+    if (toIndex) {
+      for (let a = 0; a < toIndex; a++)
+        this._listIgnoreRemove.push(a)
+    }
+    return this;
+  }
+
   user(userId: ID) {
     this._userId = userId;
     return this;
@@ -170,7 +240,7 @@ export default class SaveFile {
         ]
       })
     }
-    if (this._processType) {
+    if (! this._ignoreOldProcessType && this._processType) {
       filesToSave.forEach((p)=> {
         if (p.process !== this._processType) {
           throw new AppResponse(400)
@@ -181,6 +251,16 @@ export default class SaveFile {
     if (this._staticFile) {
       if (filesToSave.length !== 1) {
         throw new SystemError('SaveFile StaticFile mode expected one file!')
+      }
+      if (this._processType && this._ignoreOldProcessType) {
+        const file = filesToSave[0]!
+        const processedBufferFile: Buffer = await Processes[this._processType as Upload.ProcessType](
+          await fs.promises.readFile(path.join(UD, file.directory, file.filename))
+        )
+        fs.promises.writeFile(
+          path.join(UD, this._destinationDirectory, this._staticFile),
+          processedBufferFile
+        )
       }
     }
 
@@ -200,9 +280,8 @@ export default class SaveFile {
 
     if (filesToSave) {
       await fs.promises.mkdir(path.join(UD, this._destinationDirectory), { recursive: true });
-
       await Promise.allSettled(
-        filesToSave.map((p)=> new Promise<Upload | undefined>(async(resolve, reject)=> {
+        filesToSave.map((p, index)=> new Promise<Upload | undefined>(async(resolve, reject)=> {
           try {
             const fileName = this._staticFile || p.filename;
             if (! fs.existsSync(path.join(UD, p.directory, p.filename))) return reject(new SystemError(`File not exists: ${path.join(UD, p.directory, p.filename)}`))
@@ -210,7 +289,9 @@ export default class SaveFile {
               path.join(UD, p.directory, p.filename),
               path.join(UD, this._destinationDirectory, fileName),
             )
-            await fs.promises.rm(path.join(UD, p.directory, p.filename))
+            if (! this._listIgnoreRemove.includes(index)) {
+              await fs.promises.rm(path.join(UD, p.directory, p.filename))
+            }
             p.directory = this._destinationDirectory
             p.destination = path.join(UD, this._destinationDirectory)
             await p.save({session: this._session})
